# Make `QueryString{Encoder,Decoder}` even faster