Motivation:

@nitsanw and @franz1981 came up with an even faster way to convert a
character into a primitive value:

- Allocate an array as long as 65536.
- Use a character as an index.
  - No need for boundary check at all because a `char` never exceeds
    65535 by definition.
  - Increased cache miss may be a concern, but I guess it will not
    matter because most characters will be in ASCII range anyway.

Modifications:

- Use a 65536-byte long lookup table for decoding a hexdecimal nibble
  and determining if a character needs to be percent-encoded.

Result:

- Nice performance boost