@trustin 2019-12-21
Before (unsafe array access):
```
Benchmark                                    Mode  Cnt       Score      Error  Units
QueryStringDecoderBenchmark.armeriaAscii    thrpt   50  377588.237 ± 5587.696  ops/s
QueryStringDecoderBenchmark.armeriaLong     thrpt   50   19582.624 ±  457.648  ops/s
QueryStringDecoderBenchmark.armeriaMixed    thrpt   50  119342.608 ± 2209.684  ops/s
QueryStringDecoderBenchmark.armeriaUnicode  thrpt   50   86064.748 ± 2097.511  ops/s
QueryStringEncoderBenchmark.armeriaAscii    thrpt   50  844493.147 ± 1773.086  ops/s
QueryStringEncoderBenchmark.armeriaLong     thrpt   50   21028.206 ±  927.736  ops/s
QueryStringEncoderBenchmark.armeriaMixed    thrpt   50  226818.284 ± 7212.655  ops/s
QueryStringEncoderBenchmark.armeriaUnicode  thrpt   50  216179.147 ± 2157.503  ops/s
```

After (64k lookup table):
```
Benchmark                                    Mode  Cnt       Score      Error  Units
QueryStringDecoderBenchmark.armeriaAscii    thrpt   50  385508.751 ± 5165.823  ops/s
QueryStringDecoderBenchmark.armeriaLong     thrpt   50   20252.521 ±  434.238  ops/s
QueryStringDecoderBenchmark.armeriaMixed    thrpt   50  129201.709 ± 4710.201  ops/s
QueryStringDecoderBenchmark.armeriaUnicode  thrpt   50   88696.118 ±  418.647  ops/s
QueryStringEncoderBenchmark.armeriaAscii    thrpt   50  890959.585 ± 7127.986  ops/s
QueryStringEncoderBenchmark.armeriaLong     thrpt   50   22346.938 ±  527.001  ops/s
QueryStringEncoderBenchmark.armeriaMixed    thrpt   50  227038.288 ±  926.122  ops/s
QueryStringEncoderBenchmark.armeriaUnicode  thrpt   50  212527.522 ± 2766.579  ops/s
```

Notes:
- `Decoder.armeriaAscii()` shouldn't be affected by this change.
- `Encoder.armeriaAscii()` should be most-affected by this change.

The result is somewhat noisy but `Decoder.armeriaUnicode()` and `Encoder.armeriaAscii()` show obvious win, so I think we can say this change improves throughput.

What's interesting is that we lost for `Encoder.armeriaUnicode()`, which might be due to increased cache miss.