@franz1981 2019-12-28
@anuraaga any news re https://github.com/line/armeria/pull/2336#discussion_r360785367? I still don't know if you have found that we have unwanted bounds checks somewhere: in my experiments it doesn't seem the case....