@anuraaga 2019-12-29
@franz1981 I haven't done any more investigation and am not planning to right now since not sure what else to try :)

I was just printing assembly for the code snippet I pasted, which matches the pattern for hex decide, and saw the branch. This doesn't surprise me since bytecode shouldn't see any difference between int and char for local variables on the stack, everything is just an int at the bytecode level (I'm guessing the char param comment I see in the output is from the classloader trace and not the JIT). 

As in practice the bounds check will always pass, I guess it's hard to write a benchmark since the branch will always be correctly predicted. That's probably enough reason to not worry too much.